settings system changes work as expected
the changes to the settings system where tested and they worked.
the config file will have to be deleted in order for this to work
